[feat] pickerView 파일 분리, 텍스트 랜덤하게 출력되도록 변경
